+++ gridsome/lib/pages/__tests__/pages.spec.js
@@ -283,6 +283,77 @@ test('remove pages by component', async () => {
   expect(pages.pages()).toHaveLength(1)
 })
 
+test('find and remove pages by query', async () => {
+  const { pages } = await createApp()
+
+  pages.createPage({
+    path: '/page',
+    component: './__fixtures__/DefaultPage.vue'
+  })
+
+  expect(pages.pages()).toHaveLength(2)
+
+  pages.findAndRemovePages({ path: { '$eq': '/page' }})
+
+  expect(pages.pages()).toHaveLength(1)
+} )
+
+test('find pages by query', async () => {
+  const { pages } = await createApp()
+
+  const postSlugs = [
+    'lorem-ipsum-dolor-amet-brunch',
+    'tofu-schlitz-knausgaard-lomo',
+    'vaporware-dreamcatcher-tousled',
+    'godard-ramps-butcher-mumblecore',
+    'listicle-tattooed-quinoa-poke-occupy'
+  ]
+
+  for (const slug of postSlugs) {
+    pages.createPage({
+      path: `/posts/${slug}`,
+      component: './__fixtures__/DefaultPage.vue'
+    })
+  }
+
+
+  expect(pages.pages()).toHaveLength(6)
+
+  let matchingPages = pages.findPages({ path: { '$eq': '/posts/lorem-ipsum-dolor-amet-brunch' }})
+  expect(matchingPages).toHaveLength(1)
+
+  matchingPages = pages.findPages({ path: { '$contains': 'posts' }})
+  expect(matchingPages).toHaveLength(5)
+})
+
+test('find page by query', async () => {
+  const { pages } = await createApp()
+
+  const postSlugs = [
+    'lorem-ipsum-dolor-amet-brunch',
+    'tofu-schlitz-knausgaard-lomo',
+    'vaporware-dreamcatcher-tousled',
+    'godard-ramps-butcher-mumblecore',
+    'listicle-tattooed-quinoa-poke-occupy'
+  ]
+
+  for (const slug of postSlugs) {
+    pages.createPage({
+      path: `/posts/${slug}`,
+      component: './__fixtures__/DefaultPage.vue'
+    })
+  }
+
+  expect(pages.pages()).toHaveLength(6)
+
+  const pathToMatch = '/posts/lorem-ipsum-dolor-amet-brunch'
+  const matchingPage = pages.findPage({ path: { '$eq': pathToMatch } })
+
+  expect(matchingPage).toEqual(
+    expect.objectContaining({ path: pathToMatch })
+  )
+})
+
 test('api.createManagedPages() should only be called once', async () => {
   const createPages = jest.fn()
   const createManagedPages = jest.fn()
@@ -456,7 +527,7 @@ describe('dynamic pages', () => {
     expect(route.path).toEqual('/user/:id')
     expect(route.id).toEqual('9c1d306d222b94fa197459d7b9a32712')
     expect(route.options.name).toEqual('__user_id')
-    expect(route.internal.regexp).toEqual(/^\/user\/([^/]+?)(?:\/)?$/i)
+    expect(route.internal.regexp).toEqual(/^\/user\/([^\/]+?)(?:\/)?$/i) // eslint-disable-line no-useless-escape
     expect(route.internal.isDynamic).toEqual(true)
   })
 })
@@ -477,7 +548,7 @@ describe('create routes', () => {
     expect(route.path).toEqual('/page/:id/:page(\\d+)?/')
     expect(route.internal.path).toEqual('/page/:id')
     expect(route.internal.isDynamic).toEqual(true)
-    expect(route.internal.regexp).toEqual(/^\/page\/([^/]+?)(?:\/(\d+))?(?:\/)?$/i)
+    expect(route.internal.regexp).toEqual(/^\/page\/([^\/]+?)(?:\/(\d+))?(?:\/)?$/i) // eslint-disable-line no-useless-escape
   })
 
   test('add pages to route', async () => {